+++ src/route-handlers/photos.mjs
