+++ src/www/components/pages/contact-page.jsx
@@ -2,9 +2,9 @@ import React from "react";
 import ReactMarkdown from "react-markdown";
 import remarkGfm from "remark-gfm";
 import rehypeRaw from "rehype-raw";
-import text from "../../../content/contact-page/contact-page.md";
-import textEn from "../../../content/contact-page/contact-page-en.md";
-import { isEnglish } from "../util";
+import text from "/content/contact-page/contact-page.md";
+import textEn from "/content/contact-page/contact-page-en.md";
+import { isEnglish } from "/src/www/util";
 
 const me = () => (
   <div className="page">
