+++ src/www/components/pages/privacy-policy-page.jsx
@@ -2,9 +2,9 @@ import React from "react";
 import ReactMarkdown from "react-markdown";
 import remarkGfm from "remark-gfm";
 import rehypeRaw from "rehype-raw";
-import text from "../../../content/privacy-policy/privacy-policy.md";
-import textEn from "../../../content/privacy-policy/privacy-policy-en.md";
-import { isEnglish } from "../util";
+import text from "/content/privacy-policy/privacy-policy.md";
+import textEn from "/content/privacy-policy/privacy-policy-en.md";
+import { isEnglish } from "/src/www/util";
 
 const me = () => (
   <div className="page">
