+++ src/www/components/widgets/newsfeed.jsx
@@ -2,7 +2,7 @@ import React from "react";
 import ReactMarkdown from "react-markdown";
 import remarkGfm from "remark-gfm";
 import rehypeRaw from "rehype-raw";
-import { isEnglish } from "../../util";
+import { isEnglish } from "/src/www/util";
 
 const fetchNews = (num) => (
     fetch(`/newsfeed?num=${num}`)
